Remove unneeded trait
